src/on_chain_analysis/on_chain_analysis_agent.py
import pandas as pd
import numpy as np
import requests
from bs4 import BeautifulSoup
from sqlalchemy import create_engine
import logging

# Set up logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

class AdvancedOnChainAnalysisAgent:
    def __init__(self, db_url):
        """
        Initialize the AdvancedOnChainAnalysisAgent.

        :param db_url: Database URL (e.g., 'sqlite:///data.db').
        """
        self.db_url = db_url
        self.engine = create_engine(db_url)

    def load_data(self, table_name):
        """
        Load data from the database.

        :param table_name: Name of the table to load data from.
        :return: DataFrame containing the data.
        """
        query = f"SELECT * FROM {table_name}"
        data = pd.read_sql(query, self.engine)
        return data

    def fetch_on_chain_data(self, asset):
        """
        Fetch on-chain data for a given crypto asset.

        :param asset: Crypto asset symbol (e.g., 'BTC-USD').
        :return: DataFrame containing on-chain data.
        """
        if not asset.endswith('-USD'):
            return pd.DataFrame()

        try:
            url = f"https://api.blockchain.info/charts/transactions-per-second?timespan=5weeks&format=json"
            response = requests.get(url)
            data = response.json()
            df = pd.DataFrame(data['values'])
            df['Asset'] = asset
            return df
        except Exception as e:
            logging.error(f"Error fetching on-chain data for {asset}: {e}")
            return pd.DataFrame()

    def analyze_transaction_volume(self, data):
        """
        Analyze transaction volume trends.

        :param data: DataFrame containing on-chain data.
        :return: DataFrame with transaction volume analysis.
        """
        data['Transaction Volume Trend'] = data['y'].rolling(window=7).mean()
        return data

    def analyze_wallet_behavior(self, data):
        """
        Analyze wallet behavior (e.g., large wallet movements).

        :param data: DataFrame containing on-chain data.
        :return: DataFrame with wallet behavior analysis.
        """
        # Placeholder for wallet behavior analysis
        data['Large Wallet Movement'] = np.where(data['y'] > data['y'].quantile(0.95), 1, 0)
        return data

    def analyze_smart_contract_interactions(self, data):
        """
        Analyze smart contract interactions.

        :param data: DataFrame containing on-chain data.
        :return: DataFrame with smart contract interaction analysis.
        """
        # Placeholder for smart contract interaction analysis
        data['Smart Contract Interaction'] = np.random.randint(0, 2, data.shape[0])
        return data

    def perform_on_chain_analysis(self, table_name):
        """
        Perform on-chain analysis on the data.

        :param table_name: Name of the table to perform analysis on.
        :return: DataFrame with on-chain analysis results.
        """
        logging.info(f"Loading data from {table_name}...")
        data = self.load_data(table_name)

        if data is None or data.empty:
            logging.warning(f"No data found in {table_name}. Returning empty DataFrame.")
            return pd.DataFrame()

        logging.info("Fetching on-chain data...")
        for asset in data['Asset'].unique():
            on_chain_data = self.fetch_on_chain_data(asset)
            if not on_chain_data.empty:
                data = data.merge(on_chain_data, on='Asset', how='left')

        logging.info("Analyzing transaction volume...")
        data = self.analyze_transaction_volume(data)

        logging.info("Analyzing wallet behavior...")
        data = self.analyze_wallet_behavior(data)

        logging.info("Analyzing smart contract interactions...")
        data = self.analyze_smart_contract_interactions(data)

        return data
    def store_analysis_results(self, data, table_name):
        """
        Store the analysis results in the database.

        :param data: DataFrame containing analysis results.
        :param table_name: Name of the table to store data.
        """
        data.to_sql(table_name, self.engine, if_exists='replace', index=True)
        logging.info(f"Analysis results stored in {table_name}.")

if __name__ == "__main__":
    # Define database URL
    db_url = 'sqlite:///advanced_historical_data.db'

    # Initialize the AdvancedOnChainAnalysisAgent
    agent = AdvancedOnChainAnalysisAgent(db_url)

    # Perform on-chain analysis
    table_name = 'preprocessed_data'
    analysis_results = agent.perform_on_chain_analysis(table_name)

    # Store analysis results
    analysis_table_name = 'on_chain_analysis_results'
    agent.store_analysis_results(analysis_results, analysis_table_name)