requirements.txt
pandas==1.3.5
numpy==1.21.5
scikit-learn==1.0.2
yfinance==0.1.70
sqlalchemy==1.4.32
tensorflow==2.8.0
keras>=2.8.0rc0
scipy==1.7.3
requests==2.31.0
beautifulsoup4==4.10.0
schedule==1.1.0
pyyaml==6.0